Added History window
Added a "File" menu to the game window and added a new pop-up window with the history (ledger) of the current reign.  "Closing" the history window does not destroy, just withdraws the window.  Had to update base code of Hammurabi.py to change initialization of Ruler.bushels_to_feed from zero to what it must have been (from the last year of previous ruler.